Refactored contact method for footer
