+++ src/components/Footer/ContactMethod/ContactMethod.stories.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import ContactMethod from './ContactMethod';
+
+export default {
+  title: 'Components/Footer/ContactMethod',
+  component: ContactMethod,
+  argTypes: {
+    arrayOfExperiences: { control: 'object' },
+  },
+};
+
+const Template = args => <ContactMethod {...args}/>;
+
+export const ContactMethodStory = Template.bind({});
+ContactMethodStory.args = {
+  
+};
