+++ src/components/Footer/ContactMethod/index.js
@@ -1,0 +1,3 @@
+import ContactMethod from './ContactMethod';
+
+export default ContactMethod;
